clean: container definitions
